refactor(views/rules) - reorganize the folders into builders, adapters and fields
