+++ views/components/rules/adapters/__init__.py
@@ -1,0 +1,3 @@
+from .rule_adapter_factory import RuleAdapterFactory
+
+__all__ = [RuleAdapterFactory]
