+++ views/components/rules/builders/__init__.py
@@ -3,6 +3,7 @@ from .trigger_builder import TriggerBuilder
 from .action_trigger_details_builder import ActionTriggerDetailsBuilder
 from .conditions_builder import ConditionsBuilder
 from .actions_builder import ActionsBuilder
+from .rule_factory import RuleFactory
 
 __all__ = [
     "GeneralSettingsBuilder",
@@ -10,4 +11,5 @@ __all__ = [
     "ActionTriggerDetailsBuilder",
     "ConditionsBuilder",
     "ActionsBuilder",
+    "RuleFactory",
 ]
